requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.56.0
pip_system_certs==5.3